Update KellyGEN.py
Added capacity to hunt down and interpret adjective constructions, as well as multiple constructional elements associated with a given query via an update to both 'oblique()' and 'corpus_callosum()' elements. For adjective constructions, I haven't found a good way to label them as such, but the system replaces the 'oblique' elements in the array spat out by corpus_callosum() with the adjective phrase elements.