Fixed problem with OSError
Getting tired of the whole thing breaking down when it runs into an OSError? Or more to the point, are you tired of not being able to parse spoken text, which tends to be too long for the Stanford Parser to handle, and then having the whole system break? I was pretty tired of that as well. Created a workaround for Occipital() that excepts OSErrors in order to continue running through a corpus.